LayoutTests/media/modern-media-controls/tracks-panel/tracks-panel-hide-expected.txt

Testing a TracksPanel is hidden with an animation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


tracksPanel.presentInParent(container)
PASS tracksPanel.presented is true

tracksPanel.hide()
PASS tracksPanel.presented is false

Transition ended
PASS tracksPanel.presented is false

PASS successfullyParsed is true

TEST COMPLETE
